Alternate universe in which slow things are done via queues

+++ shakenfist/virt.py
@@ -15,7 +15,6 @@ from oslo_concurrency import processutils
 
 from shakenfist import config
 from shakenfist import db
-from shakenfist import etcd
 from shakenfist import images
 from shakenfist import net
 from shakenfist import util
@@ -142,6 +141,10 @@ class Instance(object):
 
         self.db_entry['block_devices']['finalized'] = False
 
+    # NOTE(mikal): this method is now strictly the instance specific steps for
+    # creation. It is assumed that the image sits in local cache already, and
+    # has been transcoded to the right format. This has been done to facilitate
+    # moving to a queue and task based creation mechanism.
     def create(self, lock=None):
         db.update_instance_state(self.db_entry['uuid'], 'creating')
 
@@ -152,7 +155,7 @@ class Instance(object):
             os.makedirs(self.instance_path)
 
         # Generate a config drive
-        with util.RecordedOperation('make config drive', self) as _:
+        with util.RecordedOperation('make config drive', self):
             self._make_config_drive(os.path.join(
                 self.instance_path, self.db_entry['block_devices']['devices'][1]['path']))
 
@@ -161,19 +164,16 @@ class Instance(object):
             modified_disks = []
             for disk in self.db_entry['block_devices']['devices']:
                 if disk.get('base'):
-                    try:
-                        hashed_image_path = images.get_image(disk['base'],
-                                                             [lock], self)
-                    except etcd.LockException:
-                        # TODO(andy): Tidy up the failed instance creation
-                        raise
+                    hashed_image_path = images.get_image(
+                        disk['base'], [lock], self)
 
-                    try:
-                        cd = pycdlib.PyCdlib()
-                        cd.open(hashed_image_path)
-                        disk['present_as'] = 'cdrom'
-                    except Exception:
-                        pass
+                    with util.RecordedOperation('detect cdrom images', self):
+                        try:
+                            cd = pycdlib.PyCdlib()
+                            cd.open(hashed_image_path)
+                            disk['present_as'] = 'cdrom'
+                        except Exception:
+                            pass
 
                     if disk.get('present_as', 'cdrom') == 'cdrom':
                         # There is no point in resizing or COW'ing a cdrom
@@ -197,13 +197,7 @@ class Instance(object):
                         disk['device'] = 'hd%s' % disk['device'][-1]
                         disk['bus'] = 'ide'
                     else:
-                        with util.RecordedOperation('transcode image', self) as _:
-                            if lock:
-                                lock.refresh()
-
-                            images.transcode(hashed_image_path)
-
-                        with util.RecordedOperation('resize image', self) as _:
+                        with util.RecordedOperation('resize image', self):
                             if lock:
                                 lock.refresh()
 
@@ -211,7 +205,7 @@ class Instance(object):
                                 hashed_image_path, disk['size'])
 
                         if config.parsed.get('DISK_FORMAT') == 'qcow':
-                            with util.RecordedOperation('create copy on write layer', self) as _:
+                            with util.RecordedOperation('create copy on write layer', self):
                                 if lock:
                                     lock.refresh
 
@@ -225,7 +219,7 @@ class Instance(object):
                                                '      </backingStore>' % resized_image_path)
 
                         elif config.parsed.get('DISK_FORMAT') == 'qcow_flat':
-                            with util.RecordedOperation('create flat layer', self) as _:
+                            with util.RecordedOperation('create flat layer', self):
                                 if lock:
                                     lock.refresh
 
@@ -233,7 +227,7 @@ class Instance(object):
                                     resized_image_path, disk['path'])
 
                         elif config.parsed.get('DISK_FORMAT') == 'flat':
-                            with util.RecordedOperation('create raw disk', self) as _:
+                            with util.RecordedOperation('create raw disk', self):
                                 if lock:
                                     lock.refresh
 
@@ -257,15 +251,15 @@ class Instance(object):
             self.db_entry['uuid'], self.db_entry['block_devices'])
 
         # Create the actual instance
-        with util.RecordedOperation('create domain XML', self) as _:
+        with util.RecordedOperation('create domain XML', self):
             self._create_domain_xml()
-        with util.RecordedOperation('create domain', self) as _:
+        with util.RecordedOperation('create domain', self):
             self.power_on()
 
         db.update_instance_state(self.db_entry['uuid'], 'created')
 
     def delete(self):
-        with util.RecordedOperation('delete domain', self) as _:
+        with util.RecordedOperation('delete domain', self):
             try:
                 self.power_off()
 
@@ -274,22 +268,22 @@ class Instance(object):
             except Exception as e:
                 util.ignore_exception('instance delete', e)
 
-        with util.RecordedOperation('delete disks', self) as _:
+        with util.RecordedOperation('delete disks', self):
             try:
-                shutil.rmtree(self.instance_path)
+                if os.path.exists(self.instance_path):
+                    shutil.rmtree(self.instance_path)
             except Exception as e:
                 util.ignore_exception('instance delete', e)
 
-        with util.RecordedOperation('release network addresses', self) as _:
+        with util.RecordedOperation('release network addresses', self):
             for ni in db.get_instance_interfaces(self.db_entry['uuid']):
                 db.update_network_interface_state(ni['uuid'], 'deleted')
-                with db.get_lock('sf/ipmanager/%s' % ni['network_uuid'],
-                                 ttl=120) as _:
+                with db.get_lock('ipmanager', None, ni['network_uuid'],
+                                 ttl=120):
                     ipm = db.get_ipmanager(ni['network_uuid'])
                     ipm.release(ni['ipv4'])
                     db.persist_ipmanager(ni['network_uuid'], ipm.save())
 
-        db.update_instance_state(self.db_entry['uuid'], 'deleted')
         db.update_instance_power_state(self.db_entry['uuid'], 'off')
         db.free_console_port(self.db_entry['console_port'])
         db.free_console_port(self.db_entry['vdi_port'])
@@ -542,7 +536,7 @@ class Instance(object):
             if d['type'] != 'qcow2':
                 continue
 
-            with util.RecordedOperation('snapshot %s' % d['device'], self) as _:
+            with util.RecordedOperation('snapshot %s' % d['device'], self):
                 self._snapshot_device(
                     d['path'], os.path.join(snappath, d['device']))
                 db.create_snapshot(snapshot_uuid, d['device'], self.db_entry['uuid'],
